ActiveDecoratorの導入。 user, articleテーブルに対応するdecorationファイル作成。
