+++ app/models/user.rb
@@ -52,17 +52,6 @@ class User < ApplicationRecord
     likes.exists?(article_id: article.id)
   end
 
-  def display_name
-    # if profile && profile.nickname
-    #   profile.nickname
-    # else
-    #   self.email.split('@').first
-    # end
-
-    # &.ぼっち演算子:nillでない場合、nicknameを返す
-    profile&.nickname || self.email.split('@').first
-  end
-
   # def birthday
   #   profile&.birthday
   # end
@@ -75,15 +64,6 @@ class User < ApplicationRecord
   # ない場合、新たにレコードを作成する
   def prepare_profile
     profile || build_profile
-  end
-
-  def avatar_image
-    # profileデータがある and profile.avatarがある
-    if profile&.avatar&.attached?
-      profile.avatar
-    else
-      'default-avatar.png'
-    end
   end
 
   def follow!(user)
